mapper-faas/secrets/requirements.txt
appnope==0.1.3
asttokens==2.0.8
attrs==22.1.0
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.11.1
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cmudict==1.0.2
constantly==15.1.0
cryptography==38.0.3
cssselect==1.2.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.1.1
filelock==3.8.0
firebase-admin==6.0.1
Flask==2.2.2
future==0.18.2
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.50.0
httplib2==0.21.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipykernel==6.16.0
ipython==8.5.0
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jupyter-core==4.11.1
jupyter_client==7.3.5
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
msgpack==1.0.4
nest-asyncio==1.5.6
nltk==3.7
numpy==1.23.4
oauth2client==4.1.3
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
parsel==1.7.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
Protego==0.2.1
proto-plus==1.22.1
protobuf==4.21.9
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
Pygments==2.13.0
PyJWT==2.6.0
pymemcache==3.5.2
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyspark==3.3.0
python-dateutil==2.8.2
python-twitter==3.5
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
queuelib==1.6.2
regex==2022.9.13
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
Scrapy==2.7.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.5.1
tldextract==3.4.0
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
tweepy==4.12.0
Twisted==22.10.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
w3lib==2.0.1
wcwidth==0.2.5
Werkzeug==2.2.2
wikipedia==1.4.0
zope.interface==5.5.2
